데이터 파이프라인 테스트 코드 구현 완료
