+++ pytest.ini
@@ -1,0 +1,40 @@
+[pytest]
+# pytest 설정 파일
+
+# 테스트 경로
+testpaths = tests
+
+# Python 파일 패턴
+python_files = test_*.py
+
+# Python 클래스 패턴
+python_classes = Test*
+
+# Python 함수 패턴
+python_functions = test_*
+
+# 마커 정의
+markers =
+    slow: 느린 테스트 (실행에 5초 이상 소요)
+    integration: 통합 테스트
+    unit: 단위 테스트
+    ml: 머신러닝 관련 테스트
+
+# 기본 옵션
+addopts =
+    -ra
+    --strict-markers
+    --ignore=tests/integration/test_run_pipeline.py
+    --ignore=tests/unit/test_all_ticker.py
+    --ignore=tests/integration/test_train_model.py
+
+# 로깅 설정
+log_cli = false
+log_cli_level = WARNING
+
+# 경고 필터
+filterwarnings =
+    ignore::DeprecationWarning:pandas_ta_classic.*
+
+# 커버리지 설정 (pytest-cov 사용 시)
+# --cov=src --cov-report=html --cov-report=term-missing
